+++ src/apps/ENGINE/src/data.cpp
@@ -272,7 +272,7 @@ void DATA::SetPtr(uintptr_t value)
     pValue = value;
 }
 
-void DATA::Set(long value)
+void DATA::Set(int32_t value)
 {
     // if(bRef)
     if (Data_type == VAR_REFERENCE)
@@ -484,7 +484,7 @@ bool DATA::GetPtr(uintptr_t &value)
     return false;
 }
 
-bool DATA::Get(long &value)
+bool DATA::Get(int32_t &value)
 {
     // if(bRef)
     if (Data_type == VAR_REFERENCE)
@@ -587,7 +587,7 @@ bool DATA::Get(const char *attribute_name, const char *&value)
     return true;
 }
 
-bool DATA::Get(long &value, uint32_t index)
+bool DATA::Get(int32_t &value, uint32_t index)
 {
     // if(bRef)
     if (Data_type == VAR_REFERENCE)
@@ -697,7 +697,7 @@ bool DATA::Get(const char *&value, uint32_t index)
       return true;    */
 }
 
-bool DATA::Set(long value, uint32_t index)
+bool DATA::Set(int32_t value, uint32_t index)
 {
     // if(bRef)
     if (Data_type == VAR_REFERENCE)
@@ -1328,7 +1328,7 @@ bool DATA::Inverse()
 
 bool DATA::Power(DATA *pV)
 {
-    long lV;
+    int32_t lV;
     if (pV->GetType() != VAR_INTEGER)
     {
         Error("bad power argument");
@@ -1338,7 +1338,7 @@ bool DATA::Power(DATA *pV)
     return Power(lV);
 }
 
-bool DATA::Power(long Deg)
+bool DATA::Power(int32_t Deg)
 {
     // if(bRef)
     if (Data_type == VAR_REFERENCE)
@@ -1353,8 +1353,8 @@ bool DATA::Power(long Deg)
     if (IsArray())
         return false;
 
-    long n;
-    long lBase;
+    int32_t n;
+    int32_t lBase;
     float fBase;
 
     if (IsArray())
@@ -2426,7 +2426,7 @@ bool DATA::CompareAndSetResult(DATA *pV, S_TOKEN_TYPE op)
     pV = pV->GetVarPointer();
     if (pV == nullptr)
     {
-        Set(static_cast<long>(0));
+        Set(0);
         return false;
     }
 
@@ -2434,7 +2434,7 @@ bool DATA::CompareAndSetResult(DATA *pV, S_TOKEN_TYPE op)
     {
         if (!Convert(VAR_INTEGER))
         {
-            Set(static_cast<long>(0));
+            Set(0);
             return false;
         }
     }
@@ -2443,7 +2443,7 @@ bool DATA::CompareAndSetResult(DATA *pV, S_TOKEN_TYPE op)
     {
         if (!Convert(VAR_FLOAT))
         {
-            Set(static_cast<long>(0));
+            Set(0);
             return false;
         }
     }
@@ -2462,51 +2462,51 @@ bool DATA::CompareAndSetResult(DATA *pV, S_TOKEN_TYPE op)
                 pV->Convert(VAR_INTEGER);
             case OP_BOOL_EQUAL:
                 if (lValue == pV->lValue)
-                    Set(static_cast<long>(1));
+                    Set(1);
                 else
-                    Set(static_cast<long>(0));
+                    Set(0);
                 break;
             case OP_GREATER:
                 if (lValue > pV->lValue)
-                    Set(static_cast<long>(1));
+                    Set(1);
                 else
-                    Set(static_cast<long>(0));
+                    Set(0);
                 break;
             case OP_GREATER_OR_EQUAL:
                 if (lValue >= pV->lValue)
-                    Set(static_cast<long>(1));
+                    Set(1);
                 else
-                    Set(static_cast<long>(0));
+                    Set(0);
                 break;
             case OP_LESSER:
                 if (lValue < pV->lValue)
-                    Set(static_cast<long>(1));
+                    Set(1);
                 else
-                    Set(static_cast<long>(0));
+                    Set(0);
                 break;
             case OP_LESSER_OR_EQUAL:
                 if (lValue <= pV->lValue)
-                    Set(static_cast<long>(1));
+                    Set(1);
                 else
-                    Set(static_cast<long>(0));
+                    Set(0);
                 break;
             case OP_NOT_EQUAL:
                 if (lValue != pV->lValue)
-                    Set(static_cast<long>(1));
+                    Set(1);
                 else
-                    Set(static_cast<long>(0));
+                    Set(0);
                 break;
             case OP_BOOL_AND:
                 if (lValue != 0 && pV->lValue != 0)
-                    Set(static_cast<long>(1));
+                    Set(1);
                 else
-                    Set(static_cast<long>(0));
+                    Set(0);
                 break;
             case OP_BOOL_OR:
                 if (lValue != 0 || pV->lValue != 0)
-                    Set(static_cast<long>(1));
+                    Set(1);
                 else
-                    Set(static_cast<long>(0));
+                    Set(0);
                 break;
             }
             break;
@@ -2516,56 +2516,56 @@ bool DATA::CompareAndSetResult(DATA *pV, S_TOKEN_TYPE op)
             {
             case OP_BOOL_EQUAL:
                 if (lValue == pV->fValue)
-                    Set(static_cast<long>(1));
+                    Set(1);
                 else
-                    Set(static_cast<long>(0));
+                    Set(0);
                 break;
             case OP_GREATER:
                 if (lValue > pV->fValue)
-                    Set(static_cast<long>(1));
+                    Set(1);
                 else
-                    Set(static_cast<long>(0));
+                    Set(0);
                 break;
             case OP_GREATER_OR_EQUAL:
                 if (lValue >= pV->fValue)
-                    Set(static_cast<long>(1));
+                    Set(1);
                 else
-                    Set(static_cast<long>(0));
+                    Set(0);
                 break;
             case OP_LESSER:
                 if (lValue < pV->fValue)
-                    Set(static_cast<long>(1));
+                    Set(1);
                 else
-                    Set(static_cast<long>(0));
+                    Set(0);
                 break;
             case OP_LESSER_OR_EQUAL:
                 if (lValue <= pV->fValue)
-                    Set(static_cast<long>(1));
+                    Set(1);
                 else
-                    Set(static_cast<long>(0));
+                    Set(0);
                 break;
             case OP_NOT_EQUAL:
                 if (lValue != pV->fValue)
-                    Set(static_cast<long>(1));
+                    Set(1);
                 else
-                    Set(static_cast<long>(0));
+                    Set(0);
                 break;
             case OP_BOOL_AND:
                 if (lValue != 0 && pV->fValue != 0)
-                    Set(static_cast<long>(1));
+                    Set(1);
                 else
-                    Set(static_cast<long>(0));
+                    Set(0);
                 break;
             case OP_BOOL_OR:
                 if (lValue != 0 || pV->fValue != 0)
-                    Set(static_cast<long>(1));
+                    Set(1);
                 else
-                    Set(static_cast<long>(0));
+                    Set(0);
                 break;
             }
             break;
         default:
-            Set(static_cast<long>(0));
+            Set(0);
             return false;
         }
         break;
@@ -2577,51 +2577,51 @@ bool DATA::CompareAndSetResult(DATA *pV, S_TOKEN_TYPE op)
             {
             case OP_BOOL_EQUAL:
                 if (fValue == pV->lValue)
-                    Set(static_cast<long>(1));
+                    Set(1);
                 else
-                    Set(static_cast<long>(0));
+                    Set(0);
                 break;
             case OP_GREATER:
                 if (fValue > pV->lValue)
-                    Set(static_cast<long>(1));
+                    Set(1);
                 else
-                    Set(static_cast<long>(0));
+                    Set(0);
                 break;
             case OP_GREATER_OR_EQUAL:
                 if (fValue >= pV->lValue)
-                    Set(static_cast<long>(1));
+                    Set(1);
                 else
-                    Set(static_cast<long>(0));
+                    Set(0);
                 break;
             case OP_LESSER:
                 if (fValue < pV->lValue)
-                    Set(static_cast<long>(1));
+                    Set(1);
                 else
-                    Set(static_cast<long>(0));
+                    Set(0);
                 break;
             case OP_LESSER_OR_EQUAL:
                 if (fValue <= pV->lValue)
-                    Set(static_cast<long>(1));
+                    Set(1);
                 else
-                    Set(static_cast<long>(0));
+                    Set(0);
                 break;
             case OP_NOT_EQUAL:
                 if (fValue != pV->lValue)
-                    Set(static_cast<long>(1));
+                    Set(1);
                 else
-                    Set(static_cast<long>(0));
+                    Set(0);
                 break;
             case OP_BOOL_AND:
                 if (fValue != 0 && pV->lValue != 0)
-                    Set(static_cast<long>(1));
+                    Set(1);
                 else
-                    Set(static_cast<long>(0));
+                    Set(0);
                 break;
             case OP_BOOL_OR:
                 if (fValue != 0 || pV->lValue != 0)
-                    Set(static_cast<long>(1));
+                    Set(1);
                 else
-                    Set(static_cast<long>(0));
+                    Set(0);
                 break;
             }
             break;
@@ -2630,56 +2630,56 @@ bool DATA::CompareAndSetResult(DATA *pV, S_TOKEN_TYPE op)
             {
             case OP_BOOL_EQUAL:
                 if (fValue == pV->fValue)
-                    Set(static_cast<long>(1));
+                    Set(1);
                 else
-                    Set(static_cast<long>(0));
+                    Set(0);
                 break;
             case OP_GREATER:
                 if (fValue > pV->fValue)
-                    Set(static_cast<long>(1));
+                    Set(1);
                 else
-                    Set(static_cast<long>(0));
+                    Set(0);
                 break;
             case OP_GREATER_OR_EQUAL:
                 if (fValue >= pV->fValue)
-                    Set(static_cast<long>(1));
+                    Set(1);
                 else
-                    Set(static_cast<long>(0));
+                    Set(0);
                 break;
             case OP_LESSER:
                 if (fValue < pV->fValue)
-                    Set(static_cast<long>(1));
+                    Set(1);
                 else
-                    Set(static_cast<long>(0));
+                    Set(0);
                 break;
             case OP_LESSER_OR_EQUAL:
                 if (fValue <= pV->fValue)
-                    Set(static_cast<long>(1));
+                    Set(1);
                 else
-                    Set(static_cast<long>(0));
+                    Set(0);
                 break;
             case OP_NOT_EQUAL:
                 if (fValue != pV->fValue)
-                    Set(static_cast<long>(1));
+                    Set(1);
                 else
-                    Set(static_cast<long>(0));
+                    Set(0);
                 break;
             case OP_BOOL_AND:
                 if (fValue != 0 && pV->fValue != 0)
-                    Set(static_cast<long>(1));
+                    Set(1);
                 else
-                    Set(static_cast<long>(0));
+                    Set(0);
                 break;
             case OP_BOOL_OR:
                 if (fValue != 0 || pV->fValue != 0)
-                    Set(static_cast<long>(1));
+                    Set(1);
                 else
-                    Set(static_cast<long>(0));
+                    Set(0);
                 break;
             }
             break;
         default:
-            Set(static_cast<long>(0));
+            Set(0);
             return false;
         }
         break;
@@ -2690,22 +2690,22 @@ bool DATA::CompareAndSetResult(DATA *pV, S_TOKEN_TYPE op)
             switch (op)
             {
             case OP_BOOL_EQUAL:
-                Set(static_cast<long>(storm::iEquals(sValue, pV->sValue)));
+                Set(storm::iEquals(sValue, pV->sValue));
                 break;
             case OP_GREATER:
-                Set(static_cast<long>(storm::iGreater(sValue, pV->sValue)));
+                Set(storm::iGreater(sValue, pV->sValue));
                 break;
             case OP_GREATER_OR_EQUAL:
-                Set(static_cast<long>(storm::iGreaterOrEqual(sValue, pV->sValue)));
+                Set(storm::iGreaterOrEqual(sValue, pV->sValue));
                 break;
             case OP_LESSER:
-                Set(static_cast<long>(storm::iLess(sValue, pV->sValue)));
+                Set(storm::iLess(sValue, pV->sValue));
                 break;
             case OP_LESSER_OR_EQUAL:
-                Set(static_cast<long>(storm::iLessOrEqual(sValue, pV->sValue)));
+                Set(storm::iLessOrEqual(sValue, pV->sValue));
                 break;
             case OP_NOT_EQUAL:
-                Set(static_cast<long>(!storm::iEquals(sValue, pV->sValue)));
+                Set(!storm::iEquals(sValue, pV->sValue));
                 break;
             case OP_BOOL_AND:
             case OP_BOOL_OR:
@@ -2715,7 +2715,7 @@ bool DATA::CompareAndSetResult(DATA *pV, S_TOKEN_TYPE op)
 
             break;
         default:
-            Set(static_cast<long>(0));
+            Set(0);
             return false;
         }
         break;
@@ -2743,24 +2743,24 @@ bool DATA::BoolConvert()
     {
     case VAR_INTEGER:
         if (lValue != 0)
-            Set(static_cast<long>(1));
+            Set(1);
         else
-            Set(static_cast<long>(0));
+            Set(0);
         break;
     case VAR_FLOAT:
         if (fValue != 0)
-            Set(static_cast<long>(1));
+            Set(1);
         else
-            Set(static_cast<long>(0));
+            Set(0);
         break;
     case VAR_STRING:
-        Set(static_cast<long>(!sValue.empty()));
+        Set(!sValue.empty());
         break;
     case VAR_PTR:
         if (pValue != 0)
-            Set(static_cast<long>(1));
+            Set(1);
         else
-            Set(static_cast<long>(0));
+            Set(0);
         break;
     default:
 
@@ -2798,7 +2798,7 @@ void DATA::BadIndex(uint32_t index, uint32_t array_size)
     Error(buffer);
 }
 
-long DATA::GetLong()
+int32_t DATA::GetLong()
 {
     return lValue;
 }
