+++ src/libs/battle_interface/src/battle_command.cpp
@@ -100,7 +100,7 @@ long BICommandList::ExecuteConfirm()
     if (m_nSelectedCommandIndex >= m_aUsedCommand.size())
         return 0; // error!
 
-    long endCode = 0;
+    int32_t endCode = 0;
     long nTargIndex = 0;
     std::string sLocName;
 
@@ -132,7 +132,7 @@ long BICommandList::ExecuteConfirm()
         m_sCurrentCommandName = "";
         break;
     default:
-        Update(m_LeftTopPoint.y, m_nCurrentCommandCharacterIndex, endCode);
+        Update(m_LeftTopPoint.y, m_nCurrentCommandCharacterIndex, static_cast<long>(endCode));
     }
     return endCode;
 }
